Reverted back changes for subscription retrieval
