+++ src/WorkflowCore/Interface/IActivityController.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace WorkflowCore.Interface
@@ -16,8 +15,6 @@ namespace WorkflowCore.Interface
     public interface IActivityController
     {
         Task<PendingActivity> GetPendingActivity(string activityName, string workerId, TimeSpan? timeout = null);
-        Task<PendingActivity> GetPendingActivity(string activityName, string workerId, CancellationToken cancellationToken = default);
-        Task<PendingActivity> GetFirstPendingActivity(string workflowId, string activityToExclude = "", CancellationToken cancellationToken = default);
         Task ReleaseActivityToken(string token);
         Task SubmitActivitySuccess(string token, object result);
         Task SubmitActivityFailure(string token, object result);
