+++ src/WorkflowCore/Interface/Persistence/ISubscriptionRepository.cs
@@ -18,8 +18,6 @@ namespace WorkflowCore.Interface
 
         Task<EventSubscription> GetFirstOpenSubscription(string eventName, string eventKey, DateTime asOf, CancellationToken cancellationToken = default);
         
-        Task<EventSubscription> GetNextWorkflowSubscription(string eventName, string workflowId, string activityToExclude, DateTime asOf, CancellationToken cancellationToken = default);
-        
         Task<bool> SetSubscriptionToken(string eventSubscriptionId, string token, string workerId, DateTime expiry, CancellationToken cancellationToken = default);
         
         Task ClearSubscriptionToken(string eventSubscriptionId, string token, CancellationToken cancellationToken = default);
